// src/app/[locale]/leads/(components)/lead.interface.ts
export interface ILead {
  id: number
  name: string
  email: string
  phone: string
  address: string
  areaId: number
  birthdate: string
  educationalQualificationDescriptionId: number
  educationalQualificationTypeId: number
  educationalQualificationIssuerId: number
  studentSourceId: number
  studentId: number
  referenceId: string
  campaignId: number
  status: string
  createdAt: string
  courseId: number
  courseName: string
}